+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 -   Add login interface [#13](https://github.com/archesproject/arches-lingo/issues/13)
 -   Add front-end router [#11](https://github.com/archesproject/arches-lingo/issues/11)
+-   Add concept and scheme serializers [#103](https://github.com/archesproject/arches-lingo/issues/103)
 -   Add backend for search [#67](https://github.com/archesproject/arches-lingo/issues/67)
 -   Add concept and scheme pages [#15](https://github.com/archesproject/arches-lingo/issues/15)
 -   Add concept hierarchy component [#18](https://github.com/archesproject/arches-lingo/issues/18)